requirements.txt
neat-python
krpc
numpy
pyautogui
pynput